jaorm-lombok/src/main/java/io/github/ulisse1996/jaorm/external/support/mock/MockName.java
package io.github.ulisse1996.jaorm.external.support.mock;

import javax.lang.model.element.Name;

public class MockName implements Name {

    private final String name;

    public MockName(String name) {
        this.name = name;
    }

    @Override
    public boolean contentEquals(CharSequence cs) {
        return name.contentEquals(cs);
    }

    @Override
    public int length() {
        return name.length();
    }

    @Override
    public char charAt(int index) {
        return name.charAt(index);
    }

    @Override
    public CharSequence subSequence(int start, int end) {
        return name.subSequence(start, end);
    }

    @Override
    public String toString() {
        return name;
    }
}
